Move mixin classes to mixins.py
